+++ src/SplinePositionInterpolator2D.cpp
@@ -109,7 +109,7 @@ void SplinePositionInterpolator2D::SFValue::update() {
   vector< H3DFloat > F2;
 
 
-  if( key_index >= 0 && (key_index + 1) < key_value.size() ) {
+  if( key_index >= 0 && (key_index + 1) < (int)key_value.size() ) {
 
     int tMinus1 = (key_size-2);
     int tN = 1;
